- rc: move backups to different host now that we have migrated
changed the name of our backup host, it was a totally different board
and due to the field dynamics, needed to co-exist for a short time, so a
different name was best.  the old one could potentially be resurrected
one day too and have to run concurrently.

(the new name is cooler anyhow...)