- rc: baksrv: switch to different backup host as the arm64 is down
oizys appears to be fried.  haven't yet tried rebooting with the I/O HAT
disconnected, that's the one that seems to have a reset issue but it's
not clear.  only thing left to try but we may or may not ever do that,
there was nothing interesting besides the data disks which are extant
and moved to other server